refactor: consolidate duplicated TestRepo across unit tests
Four identical `TestRepo` structs in `src/config/` and a duplicated
`configure_test_identity` helper in `src/git/recover.rs` are replaced
by a shared `src/testutil.rs` module. Also documents the intentional
separation between benchmark and test environment isolation.
