+++ src/config/expansion.rs
@@ -537,26 +537,7 @@ mod tests {
     use insta::assert_snapshot;
 
     use super::*;
-    use crate::shell_exec::Cmd;
-
-    /// Test fixture that creates a real temporary git repository.
-    struct TestRepo {
-        _dir: tempfile::TempDir,
-        repo: Repository,
-    }
-
-    impl TestRepo {
-        fn new() -> Self {
-            let dir = tempfile::tempdir().unwrap();
-            Cmd::new("git")
-                .args(["init"])
-                .current_dir(dir.path())
-                .run()
-                .unwrap();
-            let repo = Repository::at(dir.path()).unwrap();
-            Self { _dir: dir, repo }
-        }
-    }
+    use crate::testutil::TestRepo;
 
     fn test_repo() -> TestRepo {
         TestRepo::new()
@@ -1162,12 +1143,12 @@ mod tests {
         // Set vars data via git config
         std::process::Command::new("git")
             .args(["config", "worktrunk.state.main.vars.env", "staging"])
-            .current_dir(test._dir.path())
+            .current_dir(test.path())
             .status()
             .unwrap();
         std::process::Command::new("git")
             .args(["config", "worktrunk.state.main.vars.port", "3000"])
-            .current_dir(test._dir.path())
+            .current_dir(test.path())
             .status()
             .unwrap();
 
@@ -1222,7 +1203,7 @@ mod tests {
                 "worktrunk.state.main.vars.config",
                 r#"{"port": 3000, "debug": true}"#,
             ])
-            .current_dir(test._dir.path())
+            .current_dir(test.path())
             .status()
             .unwrap();
 
@@ -1233,14 +1214,14 @@ mod tests {
                 "worktrunk.state.main.vars.tags",
                 r#"["alpha", "beta"]"#,
             ])
-            .current_dir(test._dir.path())
+            .current_dir(test.path())
             .status()
             .unwrap();
 
         // Store a plain string (not JSON)
         std::process::Command::new("git")
             .args(["config", "worktrunk.state.main.vars.env", "staging"])
-            .current_dir(test._dir.path())
+            .current_dir(test.path())
             .status()
             .unwrap();
 
@@ -1293,7 +1274,7 @@ mod tests {
                 "worktrunk.state.main.vars.config",
                 r#"{"name": "my project", "cmd": "echo hello"}"#,
             ])
-            .current_dir(test._dir.path())
+            .current_dir(test.path())
             .status()
             .unwrap();
