+++ src/config/mod.rs
@@ -112,27 +112,7 @@ mod tests {
     use insta::assert_snapshot;
 
     use super::*;
-    use crate::git::Repository;
-    use crate::shell_exec::Cmd;
-
-    /// Test fixture that creates a real temporary git repository.
-    struct TestRepo {
-        _dir: tempfile::TempDir,
-        repo: Repository,
-    }
-
-    impl TestRepo {
-        fn new() -> Self {
-            let dir = tempfile::tempdir().unwrap();
-            Cmd::new("git")
-                .args(["init"])
-                .current_dir(dir.path())
-                .run()
-                .unwrap();
-            let repo = Repository::at(dir.path()).unwrap();
-            Self { _dir: dir, repo }
-        }
-    }
+    use crate::testutil::TestRepo;
 
     fn test_repo() -> TestRepo {
         TestRepo::new()
